Added a way to ensure that a List inside of a Simple field will be immutable
Generally speaking, Fields are supposed to be immutable (although they are not required to be so); however, collections by their nature have a tendency to break this idea, therefore I tend to wrap them with Collections.unmodifiableCollection()/List() or similar when they are truly immutable; to avoid additional checks and stuff, I rolled it into the method itself;
It's important to note, however, that there doesn't seem to be a way to check if a List is already unmodifiable to avoid a List "Matryoshka" of sorts, where an unmodifiable List is put into an unmodifiable List and so on. If this becomes an issue, a custom build unmodifiable List will be in order.