Fixed FakeUtils and added a partial update test
FixedUtils used Customer class when creating a FullCustomer; this doesn't always work, because Customer automatically sets some fields to empty if they are null (this behaviour may change in the future), which causes FullCustomer to assume certain fields were set, rather than not set.
Also, created a partial update test which tests the behaviour and required this fix.