Created Lesson API v3
This API is almost identical to API v2, however it is designed differently;
First, different classes are created for different DTO purposes;
Second, the transformation from and to these classes is handled by outside resource;
Finally, domain objects contain the logic that relates to them, not Service classes;
Overall, it is (supposedly) easier to maintain. Time will tell.